requirements.txt
playwright==1.36.0
pymysql==1.1.0
cryptography==41.0.3
python-dotenv==1.0.0